+++ package.json
@@ -10,6 +10,7 @@
     "test": "REPLICATE_API_TOKEN=test-token npm run lint && npm run build"
   },
   "dependencies": {
+    "@bytescale/upload-widget-react": "^4.10.0",
     "@headlessui/react": "^1.7.15",
     "@heroicons/react": "^2.0.18",
     "@vercel/analytics": "^1.0.1",
